ocr-ai-service/requirements.txt
# FastAPI và server
fastapi>=0.100.0
uvicorn>=0.23.0

# OCR và xử lý PDF
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
opencv-python>=4.8.0

# AI và Machine Learning
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
spacy>=3.6.0
vncorenlp>=1.0.3

# Computer Vision cho handwriting
easyocr>=1.7.0
paddleocr>=2.7.0

# Xử lý văn bản tiếng Việt
underthesea>=6.0.0
pyvi>=0.1.1

# AI tùy chọn (OpenAI backup)
openai>=1.0.0

# Xử lý dữ liệu
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6

# Validation và serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
requests>=2.28.0
joblib>=1.3.0
regex>=2023.8.8
